+++ test/hotspot.test.ts
@@ -13,8 +13,10 @@ const baseConfig: Config = {
   bugfixPatterns: [/fix/i],
   moduleDefinition: "directory",
   languages: "auto",
+  excludes: [],
   comment: true,
   generateMap: true,
+  generateArtifact: true,
   githubToken: "",
 };